添加order，permission，role，videotype Servlet，修改order、user类

+++ src/main/java/dao/OrderDao.java
@@ -17,7 +17,7 @@ public interface OrderDao {
      * @param id
      * @return
      */
-    public Order queryById(Integer id);
+    public Order queryById(Long id);
 
     /**
      * 遍历订单信息
